requirements.txt
asgiref==3.6.0
autopep8==2.0.1
beautifulsoup4==4.11.2
bs4==0.0.1
certifi==2022.12.7
charset-normalizer==3.0.1
Django==4.1.7
idna==3.4
lxml==4.9.2
pycodestyle==2.10.0
pyTelegramBotAPI==4.9.0
requests==2.28.2
soupsieve==2.4
sqlparse==0.4.3
tzdata==2022.7
urllib3==1.26.14
python-dotenv